Implement BlitPattern
Which is basically the same as BlitTemplate, just with a more restricted pattern.
Maybe its possible to implement this "more" in hardware (without the CPU having to
provide the bitmap pattern), but Trio64 only supports an 8x8 bit pattern.
So one would have to write a very specialized path for that. IMHO not worth it.