Enhance BlitRectNoMaskComplete for differing src and dest pitch
The Trio64 does not support differing pitch for blit source and destination.
If that was the case, we had to fallback to CPU default implementation.

This function is often used to save and restore small parts of the framebuffer
to/from small offscreen buffers, for instance when dragging icons around.

Implement the blit by blitting each line individually. The coordinates
in the smaller buffer have to be expressed in terms of the pitch of the larger
buffer.